Split dtype mappings out from metadata constants.
